refactor: extract conversation message composer
